+++ openair1/PHY/INIT/nr_parms.c
@@ -428,9 +428,13 @@ int nr_init_frame_parms_ue(NR_DL_FRAME_PARMS *fp,
   return 0;
 }
 
-void nr_init_frame_parms_ue_sa(NR_DL_FRAME_PARMS *frame_parms, uint64_t downlink_frequency, int32_t delta_duplex, uint8_t mu, uint16_t nr_band) {
-
-  LOG_I(PHY,"SA init parameters. DL freq %lu UL offset %d SSB numerology %d N_RB_DL %d\n",
+void nr_init_frame_parms_ue_sa(NR_DL_FRAME_PARMS *frame_parms,
+                               uint64_t downlink_frequency,
+                               int64_t delta_duplex,
+                               uint8_t mu,
+                               uint16_t nr_band)
+{
+  LOG_I(PHY,"SA init parameters. DL freq %lu UL offset %ld SSB numerology %d N_RB_DL %d\n",
         downlink_frequency,
         delta_duplex,
         mu,
@@ -465,7 +469,6 @@ void nr_init_frame_parms_ue_sa(NR_DL_FRAME_PARMS *frame_parms, uint64_t downlink
   frame_parms->samples_per_frame = 10 * frame_parms->samples_per_subframe;
 
   LOG_W(PHY, "samples_per_subframe %d/per second %d, wCP %d\n", frame_parms->samples_per_subframe, 1000*frame_parms->samples_per_subframe, frame_parms->samples_per_subframe_wCP);
-
 }
 
 void nr_dump_frame_parms(NR_DL_FRAME_PARMS *fp)
